Add tests and move constraints to separate
