+++ pyiron_lammps/compatibility/file.py
@@ -8,8 +8,8 @@ from pyiron_lammps.compatibility.calculate import (
     calc_md,
     calc_minimize,
     calc_static,
-    set_selective_dynamics,
 )
+from pyiron_lammps.compatibility.constraints import set_selective_dynamics
 from pyiron_lammps.output import parse_lammps_output
 from pyiron_lammps.potential import get_potential_by_name
 from pyiron_lammps.structure import write_lammps_datafile
